move chain-test-utils to chain-impl-mockchain
These methods are not used elsewhere so now:

- They are moved to chain-impl-mockchain.
- chain-test-utils is removed.
