+++ Cargo.toml
@@ -9,7 +9,6 @@ members = [
     "chain-crypto",
     "chain-network",
     "chain-storage",
-    "chain-test-utils",
     "chain-impl-mockchain",
     "cardano-legacy-address",
     "sparse-array",
