+++ chain-impl-mockchain/Cargo.toml
@@ -26,7 +26,6 @@ thiserror = "1.0"
 lazy_static = { version = "1.3.0", optional = true }
 cardano-legacy-address = { path= "../cardano-legacy-address" }
 rand_chacha = { version = "0.3", optional = true }
-chain-test-utils = { path = "../chain-test-utils", optional = true }
 rayon = "1.5.0"
 criterion = { version = "0.3.0", optional = true }
 rand = "0.8"
@@ -36,7 +35,6 @@ property-test-api = [
         "chain-crypto/property-test-api",
         "chain-time/property-test-api",
         "chain-addr/property-test-api",
-        "chain-test-utils",
         "quickcheck",
         "quickcheck_macros",
         "lazy_static",
@@ -48,7 +46,6 @@ with-bench = ["criterion","property-test-api"]
 quickcheck = "0.9"
 quickcheck_macros = "0.9"
 chain-core = { path = "../chain-core"}
-chain-test-utils = { path = "../chain-test-utils" }
 chain-crypto = { path = "../chain-crypto", features=["property-test-api"]}
 chain-time = { path = "../chain-time", features=["property-test-api"]}
 chain-addr = { path = "../chain-addr", features=["property-test-api"]}
