+++ chain-impl-mockchain/src/testing/mod.rs
@@ -12,3 +12,4 @@ pub use builders::*;
 pub use data::KeysDb;
 pub use gen::{TestGen, VoteTestGen};
 pub use ledger::{ConfigBuilder, LedgerBuilder, TestLedger, UtxoDb};
+pub mod serialization;
