+++ chain-impl-mockchain/src/testing/serialization.rs
