--- chain-test-utils/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "chain-test-utils"
-version = "0.1.0"
-authors = ["danielsanchezq <daniel.sanchez@iohk.io>"]
-edition = "2018"
-license = "MIT OR Apache-2.0"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-chain-core = { path = "../chain-core" }
-quickcheck = "0.9.2"
-quickcheck_macros = "0.9.1"
